requirements.txt
discord.py>=2.3,<3.0
aiohttp>=3.9,<4.0
python-dotenv>=1.0,<2.0
google-cloud-aiplatform>=1.60,<2.0
